requirements.txt
langgraph==0.0.13
groq==0.4.1
python-dotenv==1.0.*
httpx==0.27.0
flask==3.0.*